Fixed running stream problem

+++ C Code/defendCode.h
@@ -44,11 +44,11 @@ extern int isInt(char* buf, int intDigitSize);
 
 extern int promptAndGetInt();
 
-extern int intOverflow(double dnum);
+extern int intOverflow(long dnum);
 
-extern int canAdd(double dnum1, double dnum2);
+extern int canAdd(long dnum1, long dnum2);
 
-extern int canMultiply(double dnum1, double dnum2);
+extern int canMultiply(long dnum1, long dnum2);
 
 extern double getNum(char* arr);
